### 2023-06-07 #5524 @dfm
@prappleizer, @tbmiller-astro — Thanks for your submission! I'm now working to assign a handling editor and I'll keep you posted during this process.

In the meantime, please feel free to take a look at this [list of potential reviewers](https://reviewers.joss.theoj.org/reviewers) (I believe you'll need to create an account) to see if anyone would be a particularly well suited reviewer who doesn't have a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). Please don't @ the potential reviewers - just list the GitHub usernames here.
