### 2023-06-23 #5524 @eloisabentivegna
👋 @AstroBarker, @rmorgan10: could you let us know about your availability?
