### 2023-06-30 2df6505
check psf normalization and psf shape < imshape.
