+++ pysersic/rendering.py
@@ -19,6 +19,10 @@ base_profile_params =dict(
     )
 )
 
+class KernelError(Exception):
+    pass
+class PSFNormalizationWarning(Warning):
+    pass
 
 class BaseRenderer(object):
     def __init__(self, 
@@ -36,8 +40,11 @@ class BaseRenderer(object):
         """
         self.im_shape = im_shape
         self.pixel_PSF = pixel_PSF
+        if not jnp.isclose(jnp.sum(self.pixel_PSF),1.0,0.1):
+            raise PSFNormalizationWarning('PSF does not appear to be appropriately normalized; Sum(psf) is more than 0.1 away from 1.')
         self.psf_shape = jnp.shape(self.pixel_PSF)
-
+        if jnp.all(self.im_shape<self.psf_shape):
+            raise KernelError('PSF pixel image size must be smaller than science image.')
         self.x = jnp.arange(self.im_shape[0])
         self.y = jnp.arange(self.im_shape[1])
         self.X,self.Y = jnp.meshgrid(self.x,self.y)
