### 2023-06-30 fe40b7f
add exepctions file
