+++ pysersic/exceptions.py
@@ -1,0 +1,6 @@
+class RMSWarning(Warning):
+    pass
+class KernelError(Exception):
+    pass
+class PSFNormalizationWarning(Warning):
+    pass
