### 2023-06-30 7a24d3f
add MaskWarning check for overmasked images.
