+++ pysersic/pysersic.py
@@ -75,7 +75,8 @@ class BaseFitter(ABC):
             self.mask = jnp.ones_like(self.data).astype(jnp.bool_)
         else:
             self.mask = jnp.logical_not(jnp.array(mask)).astype(jnp.bool_)
-        
+            if jnp.sum(self.mask)/jnp.prod(jnp.array(self.mask.shape))>0.5:
+                raise MaskWarning('More than 50 percent of input image is masked. Is this correct? (Pysersic treats True as masked; you may need to flip your boolean array.) ')
         self.renderer = renderer(data.shape, psf, **renderer_kwargs)
     
         self.prior_dict = {}
