### 2023-06-30 d72b25d
mask uses false as mask -- all True = no masking
