### 2023-06-30 fe244fa
move checks to a check_input_data function.
