+++ pysersic/pysersic.py
@@ -59,24 +59,17 @@ class BaseFitter(ABC):
         
         self.loss_func = loss_func
 
-        if data.shape != rms.shape:
-            raise AssertionError('rms map ndims must match input data')
-            
+        
         self.data = jnp.array(data) 
         self.rms = jnp.array(rms)
-        if jnp.mean(self.rms) > 5*jnp.std(self.data):
-            raise RMSWarning('Input RMS map appears to be highly offset (>5x) in magnitude from the rms of the pixels in the input image.')
         self.psf = jnp.array(psf)
-        if not jnp.isclose(jnp.sum(self.psf),1.0,0.1):
-            raise PSFNormalizationWarning('PSF does not appear to be appropriately normalized; Sum(psf) is more than 0.1 away from 1.')
-        if jnp.all(self.data.shape<self.psf.shape):
-            raise KernelError('PSF pixel image size must be smaller than science image.')
+        
         if mask is None:
             self.mask = jnp.ones_like(self.data).astype(jnp.bool_)
         else:
             self.mask = jnp.logical_not(jnp.array(mask)).astype(jnp.bool_)
-            if jnp.sum(self.mask)/jnp.prod(jnp.array(self.mask.shape))<0.5:
-                raise MaskWarning('More than 50 percent of input image is masked. Is this correct? (Pysersic treats True/1 as masked; you may need to flip your boolean array.) ')
+        
+        check_input_data(self.data,rms=self.rms,psf=self.psf,mask=self.mask)
         self.renderer = renderer(data.shape, psf, **renderer_kwargs)
     
         self.prior_dict = {}
@@ -542,3 +535,42 @@ def train_numpyro_svi_early_stop(
 
     return SVIRunResult(svi_class.get_params(best_state), svi_state,losses)
     
+
+
+def check_input_data(data,rms,psf,mask=None):
+    """Check input data for certain conditions and raise warnings or exceptions if needed
+
+    Parameters
+    ----------
+    data : ArrayLike
+        input image (galaxy/cutout)
+    rms : ArrayLike
+       rms/error map of the data. 
+    psf : ArrayLike
+        pixelized PSF
+    mask :ArrayLike, optional
+       mask with True/1 indicating a pixel should be masked, by default None
+
+    Raises
+    ------
+    RMSWarning
+        If the rms map is highly discrepant from the rms of the data
+    PSFNormalizationWarning
+        if the pst normalization is not ~1
+    KernelError
+        if the provided PSF is larger than the input image (exception)
+    MaskWarning
+        If more than 50% of the image is masked. 
+    """
+    if data.shape != rms.shape:
+        raise ShapeMatchError('RMS map ndims must match input data ndims.')
+    if jnp.mean(rms) > 5*jnp.std(data):
+        raise RMSWarning('Input RMS map appears to be highly offset (>5x) in magnitude from the rms of the pixels in the input image.')
+    if not jnp.isclose(jnp.sum(psf),1.0,0.1):
+        raise PSFNormalizationWarning('PSF does not appear to be appropriately normalized; Sum(psf) is more than 0.1 away from 1.')
+    if jnp.all(data.shape<psf.shape):
+        raise KernelError('PSF pixel image size must be smaller than science image.')
+    if mask is not None:
+        if jnp.sum(mask)/jnp.prod(jnp.array(mask.shape))<0.5:
+            raise MaskWarning('More than 50 percent of input image is masked. Is this correct? (Pysersic treats True/1 as masked; you may need to flip your boolean array.) ')
+    return
