+++ pysersic/exceptions.py
@@ -6,3 +6,6 @@ class PSFNormalizationWarning(Warning):
     pass
 class MaskWarning(Warning):
     pass
+
+class ShapeMatchError(Exception):
+    pass
