### 2023-06-30 d34cb79
docs and also check mask shape.
