### 2023-06-30 d742200
return a bool
