### 2023-06-30 1695dd2
update docs to showcase input check
