+++ pysersic/pysersic.py
@@ -63,12 +63,7 @@ class BaseFitter(ABC):
         self.data = jnp.array(data) 
         self.rms = jnp.array(rms)
         self.psf = jnp.array(psf)
-        
-        if mask is None:
-            self.mask = jnp.ones_like(self.data).astype(jnp.bool_)
-        else:
-            self.mask = jnp.logical_not(jnp.array(mask)).astype(jnp.bool_)
-        
+        self.mask = parse_mask(mask,data)
         data_isgood = check_input_data(self.data,rms=self.rms,psf=self.psf,mask=self.mask)
         self.renderer = renderer(data.shape, psf, **renderer_kwargs)
     
@@ -536,6 +531,11 @@ def train_numpyro_svi_early_stop(
     return SVIRunResult(svi_class.get_params(best_state), svi_state,losses)
     
 
+def parse_mask(mask:ArrayLike=None,data=None):
+    if mask is None:
+        return jnp.ones_like(data).astype(jnp.bool_)
+    else:
+        return jnp.logical_not(jnp.array(mask)).astype(jnp.bool_)
 
 def check_input_data(data:ArrayLike,rms:ArrayLike,psf:ArrayLike,mask:ArrayLike=None):
     """Check input data for certain conditions and raise warnings or exceptions if needed
@@ -562,6 +562,9 @@ def check_input_data(data:ArrayLike,rms:ArrayLike,psf:ArrayLike,mask:ArrayLike=N
     MaskWarning
         If more than 50% of the image is masked. 
     """
+    rms = jnp.array(rms)
+    data = jnp.array(data)
+    psf = jnp.array(psf)
     if data.shape != rms.shape:
         raise ShapeMatchError('RMS map ndims must match input data ndims.')
     if jnp.mean(rms) > 5*jnp.std(data):
@@ -571,6 +574,7 @@ def check_input_data(data:ArrayLike,rms:ArrayLike,psf:ArrayLike,mask:ArrayLike=N
     if jnp.all(data.shape<psf.shape):
         raise KernelError('PSF pixel image size must be smaller than science image.')
     if mask is not None:
+        mask = parse_mask(mask,data)
         if jnp.sum(mask)/jnp.prod(jnp.array(mask.shape))<0.5:
             raise MaskWarning('More than 50 percent of input image is masked. Is this correct? (Pysersic treats True/1 as masked; you may need to flip your boolean array.) ')
         if mask.shape !=data.shape:
