### 2023-06-30 8cfc10d
slight hack so checkdata takes human-readable mask
