### 2023-07-07 #5524 @AstroBarker
Hi, sorry I missed this and thanks for reaching out via email. I'm happy to review. I'm familiar with the methods / etc but not the scientific context as much, but I think that is fine.
