### 2023-07-28 #5703 @editorialbot
[REVIEW]: Pysersic: A Python package for determining galaxy structural properties via Bayesian inference, accelerated with jax
**Submitting author:** @prappleizer (<a href="http://orcid.org/0000-0002-7075-9931">Imad Pasha</a>)
**Repository:** https://github.com/pysersic/pysersic
**Branch with paper.md** (empty if default branch): paper-joss
**Version:** version 0.1.3
