### 2023-07-31 b87c419
Update README.md with bibtex citation
