### 2023-08-14 #5703 @AstroBarker
Working through this now -- some initial thoughts. The paper is really well written and describes appropriately the need and novelty of the software. A couple of small things, which can be ignored: the acronyms "PPL" and "HMC" are defined, but never used again. It is fine to just remove the acronyms.
