### 2023-08-14 pysersic/pysersic#28 @AstroBarker
JOSS Review: In Progress
Hi folks! Making this issue to document my JOSS review https://github.com/openjournals/joss-reviews/issues/5703.

Working through this now -- some initial thoughts. The paper is really well written and describes appropriately the need and novelty of the software. A couple of small things, which can be ignored: the acronyms "PPL" and "HMC" are defined, but never used again. It is fine to just remove the acronyms.

I'm currently working through the [Fitting a single source](https://pysersic.readthedocs.io/en/latest/example-fit.html) example, and have run into some issues with `jax` -- when trying to construct the `FitSingle` fitter, it complains about non-numeric data:

`
TypeError: Dtype >f8 is not a valid JAX array type. Only arrays of numeric types are supported by JAX.
`
This was accomplished with the following code:
```
#!/usr/bin/env python3

import numpy as np

from pysersic.priors import autoprior
from pysersic.results import plot_image
from pysersic import check_input_data
from pysersic import FitSingle
from pysersic.loss import student_t_loss

def load_data(n):
    im = np.load(f'examp_gals/gal{n}_im.npy')
    mask = np.load(f'examp_gals/gal{n}_mask.npy')
    sig = np.load(f'examp_gals/gal{n}_sig.npy')
    psf = np.load(f'examp_gals/gal{n}_psf.npy')
    return im,mask,sig,psf

im,mask,sig,psf = load_data(2)
check_input_data(data=im,rms=sig,psf=psf,mask=mask)

#fig, ax = plot_image(im,mask,sig,psf)
#fig.savefig('out.png')

prior  = autoprior(image = im, profile_type = 'sersic', mask=mask, sky_type = 'none')

#print( prior )

fitter = FitSingle(data=im,rms=sig,mask=mask,psf=psf,prior=prior,loss_func=student_t_loss)
```
The same happens when trying to construct a `FitMulti`. Perhaps I have missed something? Or some functionality has changed with versions? The conda environment I am using is below:
`environment.yml`:
```
name: pysersic
channels:
  - conda-forge
  - defaults
dependencies:
  - python<=3.11
  - pip
  - arviz
  - asdf
  - astropy
  - corner
  - jax
  - matplotlib
  - numpy
  - numpyro
  - pandas
  - photutils
  - sep
  - scipy
  - tqdm
  - pip:
      - pysersic
```
Everything prior to constructing the fitter seemed to work as intended. Please let me know what you think.

Additional comments:
 - Please note the location of the example data in the documentation.
 - I would highly suggest including a minimally working environment that is guaranteed to work from publication time. Feel free to use the one I have created above (once any potential issues with versions are resolved).

 I will continue with my review and add additional thoughts below.
