### 2023-08-14 pysersic/pysersic#28 @prappleizer
Hi @AstroBarker! Somehow, Tim and I never ran into this, but some of our early users have; I imagine something changed in Jax. The hot fix is to set the dtype of the offending array to float32 (or 64). In this case I believe it is the psf image that for some reason is saved f8.

I'll make a PR to address this soon and coerce input images to a usable dtype within the fitter, but this fix should get you going further into the examples in the meantime.
