### 2023-08-14 91db5c8
self.psf into renderer
