### 2023-08-14 pysersic/pysersic#28 @prappleizer
The branch `dtype-coerce` now has the needed fix; feel free to check it out. It is ready for a PR as well ([see here](https://github.com/pysersic/pysersic/pull/29)) so as soon as tests have run and Tim takes a look we can merge that.
