### 2023-08-14 pysersic/pysersic#28 @AstroBarker
Ah yeah, that did the trick. Thanks for the quick turnaround!
