### 2023-08-14 pysersic/pysersic#28 @AstroBarker
A couple more minor notes. In the first code block of third example, there is a missing `import matplotlib.pyplot as plt`. Also, I noticed that the source extractor package `sep` is used. This should be listed along with the dependencies or otherwise documented somewhere. I've added it to the conda environment above.
Lastly, it would be beneficial to mention, perhaps alongside installation, that the user can test the software with `pytest` and that these tests are automatically run on all PRs.

I think that about covers it for my review... great work here! All of my points are minor and non-blocking and I'll wrap up my stuf fon the JOSS side. The only thing I feel strongly about is including, somewhere, (ideally both in the repo and the installation docs) a minimally working environment, for future-proofing. It will protect future users from potentially breaking changes that might get pushed into the dependencies.
Cheers!
