### 2023-08-15 pysersic/pysersic#28 @tbmiller-astro
Thanks @AstroBarker  These are great suggestions, working on them now. Apologies, accidentally closed the issues with the PR. Re-opening now
