### 2023-08-15 c0210eb
Mention pytest in readme and docs
