+++ setup.cfg
@@ -1,2 +1,4 @@
 [metadata]
 description-file = README.md
+name = pysersic
+version = 0.1.2
