### 2023-08-16 pysersic/pysersic#28 @JBorrow
I also had the issue with the incorrect type. Has that now been remedied?
