### 2023-08-16 pysersic/pysersic#28 @prappleizer
@JBorrow It has been remedied in the current `main` build here on GitHub, but is not propagated to `pypi` yet. I'm updating descriptions and `pyproject.toml` now and should have a new version on pip tonight, so feel free to either install from GitHub or wait until tomorrow to upgrade with pip!
