+++ docs/source/issues.rst
@@ -3,6 +3,10 @@ Issues and Contributing
 
 If you happen to find a bug, have a question or want to request a feature the best way to get in touch is to open an `issue on GitHub <https://github.com/pysersic/pysersic/issues>`_.
 
-We are happy to have anyone contribute to improving pysersic. For smaller bug fixes etc. feel free to open a pull request. For larger changes or new features it is best to first open an issue to discuss the proposed changes and implementation details. 
+We are happy to have anyone contribute to improving pysersic. For smaller bug fixes etc. feel free to open a pull request. 
 
-For anyone who is new to contributing to projects on github, a good place to start is the `astropy guide <https://docs.astropy.org/en/stable/development/workflow/development_workflow.html>`_ for a nice primer on all the thing you will need to know!
+For larger changes or new features it is best to first open an issue to discuss the proposed changes and implementation details. 
+
+When a pull request is made to the ``pysersic`` repo, the automated test suite is automatically run, and you can see if changes are consistent with the tests. You can also run the tests yourself via the ``pytest`` package: simply run ``pytest`` in the ``/tests`` directory of ``pysersic``. 
+
+For anyone who is new to contributing to projects on github, a good place to start is the `astropy guide <https://docs.astropy.org/en/stable/development/workflow/development_workflow.html>`_ for a nice primer on all the things you will need to know!
