### 2023-08-16 pysersic/pysersic#28 @prappleizer
@AstroBarker We've added two environment files, one very similar to yours above, and one which has one arbitrary but working fully-specified versions for the dependencies that can be used if users are running into issues. PR #36 adds them along with some text in the readme and readthedocs which refers users to them at the install stage. That PR also has instructions for running the test suite.
