### 2023-08-30 647a8c6
Update paper.bib
add missing doi: 10.1007/978-1-4842-2677-3_5
