+++ paper.bib
@@ -312,6 +312,7 @@ url          = {https://doi.org/10.5281/zenodo.6825092}
   author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
   year =         {2004},
   url = {https://github.com/pytest-dev/pytest},
+  doi = {10.1007/978-1-4842-2677-3_5}
 }
 @ARTICLE{Scipy:2020,
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
