### 2023-09-10 #5703 @eloisabentivegna
I've just finished reading the paper (great work!) and suggested a tiny fix. Once that has been taken care of, we can start preparing the repo for publication, following the checklist below.
