### 2023-09-11 #5703 @prappleizer
- [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. [**version 0.1.3**]
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. [https://doi.org/10.5281/zenodo.8335352]
