### 2023-09-11 #5703 @eloisabentivegna
Hi @prappleizer, the title I see on Zenodo is "pysersic v0.1.3". It should be identical to the one for this submission.
