### 2023-09-11 2b17093
ensure numpy arrays for photutils datacatalog.
