+++ pysersic/priors.py
@@ -503,9 +503,13 @@ class PySersicMultiPrior(BasePrior):
 
 class SourceProperties():
     """
-    A Class used to estimate initial guesses for source properties. If no guesses are provided, then the class will estimate them using the `photutls` package and the `data_properties()` function.
+    A Class used to estimate initial guesses for source properties. 
+    If no guesses are provided, then the class will estimate them 
+    using the `photutls` package and the `data_properties()` function.
     """
-    def __init__(self,image: np.array, mask:np.array =None):
+    def __init__(self,
+                 image: Union[np.array,jnp.array], 
+                 mask: Union[np.array,jnp.array] = None):
         """Initialize the a SourceProperties object
 
         Parameters
@@ -515,8 +519,12 @@ class SourceProperties():
         mask : np.array, optional
             pixel by pixel mask, by default None
         """
-        self.image = image 
-        self.mask = mask 
+        # Force back to numpy for photutils compatibility
+        self.image = np.array(image)
+        if mask is not None:
+            self.mask = np.array(mask) 
+        else:
+            self.mask = None
         
         if self.mask is not None:
             self.cat = data_properties(self.image,mask=self.mask.astype(bool))
@@ -540,8 +548,13 @@ class SourceProperties():
         self.set_sky_guess(**kwargs)
         return self
     
-    def set_sky_guess(self,sky_guess: Optional[float] = None,sky_guess_err:Optional[float] = None,n_pix_sample:int = 5, **kwargs)-> SourceProperties:
-        """Measure or set guess for initial sky background level. If no estimate is provided, the median of the n_pix_sample number of pixels around each edge is used
+    def set_sky_guess(self,
+                      sky_guess: Optional[float] = None,
+                      sky_guess_err:Optional[float] = None,
+                      n_pix_sample:int = 5, 
+                      **kwargs)-> SourceProperties:
+        """Measure or set guess for initial sky background level. 
+        If no estimate is provided, the median of the n_pix_sample number of pixels around each edge is used
 
         Parameters
         ----------
@@ -564,8 +577,13 @@ class SourceProperties():
             self.sky_guess_err = 2*std/np.sqrt(npix)
         return self
 
-    def set_flux_guess(self,flux_guess: Optional[float] =None,flux_guess_err: Optional[float] = None,**kwargs)-> SourceProperties:
-        """Measure or set guess for initial flux. If no estimate is provided, the flux of the source in estimated as the total flux within the sgmentatated region for the source
+    def set_flux_guess(self,
+                       flux_guess: Optional[float] = None,
+                       flux_guess_err: Optional[float] = None,
+                       **kwargs)-> SourceProperties:
+        """Measure or set guess for initial flux. 
+        If no estimate is provided, the flux of the source in estimated as the total flux 
+        within the sgmentatated region for the source
 
         Parameters
         ----------
@@ -594,8 +612,12 @@ class SourceProperties():
         self.flux_guess_err = flux_guess_err
         return self 
     
-    def set_r_eff_guess(self,r_eff_guess:Optional[float] = None, r_eff_guess_err:Optional[float] = None, **kwargs)-> SourceProperties:
-        """Measure or set guess for effective radius. If no estimate is provided, the r_eff of the source in estimated using photutils
+    def set_r_eff_guess(self,
+                        r_eff_guess:Optional[float] = None, 
+                        r_eff_guess_err:Optional[float] = None, 
+                        **kwargs)-> SourceProperties:
+        """Measure or set guess for effective radius. 
+        If no estimate is provided, the r_eff of the source in estimated using photutils
 
         Parameters
         ----------
[…]
