### 2023-09-11 9e05a77
update an actual catch for dummy image.use flag.
