### 2023-09-12 bfc3714
remove sky from individual sources
