### 2023-09-13 36fe215
use isinstance to check flag
