### 2023-09-14 #5703 @eloisabentivegna
I am observing the pdf compilation error above, but the pdf in https://github.com/openjournals/joss-papers/pull/4573 seems fine.

Is this a problem, @openjournals/aass-eics?
