+++ paper.md
@@ -30,11 +30,11 @@ The modern standard for measuring structural parameters of galaxies involves a f
 
 
 $$
-I(R) ∝ F_{\rm total} \exp \left[\left(\frac{R}{R_e}\right)^{1/n}-1\right],
+I(R) ∝ F_{\mathrm{total}} \exp \left[\left(\frac{R}{R_e}\right)^{1/n}-1\right],
 $$
 
 
-where the total flux, $F_{\rm total}$, half-light radius, $R_e$ and Sérsic index, $n$ are the parameters of interest to be fit and subsequently used to characterize a galaxy's morphology.
+where the total flux, $F_{\mathrm{total}}$, half-light radius, $R_e$ and Sérsic index, $n$ are the parameters of interest to be fit and subsequently used to characterize a galaxy's morphology.
 
 
 Here we present `pysersic`, a Bayesian framework created to facilitate the inference of structural parameters from galaxy images. It is written in pure `Python`, and built using the `jax` framework [@Bradbury:2018] allowing for just-in-time (JIT) compilation, auto-differentiation and seamless execution on CPUs, GPUs or TPUs. Inference is performed with the `numpyro` [@Phan:2019;@Bingham:2019] package utilizing gradient based methods, e.g., No U-Turn Sampling (NUTS) [@Hoffman:2014], for efficient and robust posterior estimation. `pysersic` was designed to have a user-friendly interface, allowing users to fit single or multiple sources in a few lines of code. It was also designed to scale to many images, such that it can be seamlessly integrated into current and future analysis pipelines.
@@ -50,7 +50,7 @@ Inference in `pysersic` is implemented using the `numpyro` probabilistic program
 
 # Code Description
 
-`pysersic` was designed to have a user-friendly API with sensible defaults. Tools are provided to automatically generate priors for all free parameters based on an initial characterization of a given image --- but can also easily be set manually. We provide default inference routines for NUTS MCMC and variational inference using neural flows. Users can access the underlying `numpyro` model if desired, to perform inference using any tools available within the `numpyro` ecosystem. The goal for `pysersic` is to provide a reasonable defaults for new users interested in a handful of galaxies, yet maintain the ability for advanced users to tweak options as necessary to perform inference for entire surveys.
+`pysersic` was designed to have a user-friendly API with sensible defaults. Tools are provided to automatically generate priors for all free parameters based on an initial characterization of a given image --- but can also easily be set manually. We provide default inference routines for NUTS MCMC and variational inference using neural flows. Users can access the underlying `numpyro` model if desired, to perform inference using any tools available within the `numpyro` ecosystem. The goal for `pysersic` is to provide reasonable defaults for new users interested in a handful of galaxies, yet maintain the ability for advanced users to tweak options as necessary to perform inference for entire surveys.
 
[…]
