+++ paper.bib
@@ -60,7 +60,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mowla:2019,
        author = {{Mowla}, Lamiya A. and {van Dokkum}, Pieter and {Brammer}, Gabriel B. and {Momcheva}, Ivelina and {van der Wel}, Arjen and {Whitaker}, Katherine and {Nelson}, Erica and {Bezanson}, Rachel and {Muzzin}, Adam and {Franx}, Marijn and {MacKenty}, John and {Leja}, Joel and {Kriek}, Mariska and {Marchesini}, Danilo},
         title = "{COSMOS-DASH: The Evolution of the Galaxy Size-Mass Relation since z {\ensuremath{\sim}} 3 from New Wide-field WFC3 Imaging Combined with CANDELS/3D-HST}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {galaxies: evolution, galaxies: high-redshift, galaxies: photometry, galaxies: structure, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2019,
         month = jul,
@@ -79,7 +79,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Lange:2015,
        author = {{Lange}, Rebecca and {Driver}, Simon P. and {Robotham}, Aaron S.~G. and {Kelvin}, Lee S. and {Graham}, Alister W. and {Alpaslan}, Mehmet and {Andrews}, Stephen K. and {Baldry}, Ivan K. and {Bamford}, Steven and {Bland-Hawthorn}, Joss and {Brough}, Sarah and {Cluver}, Michelle E. and {Conselice}, Christopher J. and {Davies}, Luke J.~M. and {Haeussler}, Boris and {Konstantopoulos}, Iraklis S. and {Loveday}, Jon and {Moffett}, Amanda J. and {Norberg}, Peder and {Phillipps}, Steven and {Taylor}, Edward N. and {L{\'o}pez-S{\'a}nchez}, {\'A}ngel R. and {Wilkins}, Stephen M.},
         title = "{Galaxy And Mass Assembly (GAMA): mass-size relations of z < 0.1 galaxies subdivided by S{\'e}rsic index, colour and morphology}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {galaxies: elliptical and lenticular, cD, galaxies: formation, galaxies: fundamental parameters, galaxies: spiral, galaxies: statistics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2015,
         month = mar,
@@ -99,7 +99,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Kawinwanichakij:2021,
        author = {{Kawinwanichakij}, Lalitwadee and {Silverman}, John D. and {Ding}, Xuheng and {George}, Angelo and {Damjanov}, Ivana and {Sawicki}, Marcin and {Tanaka}, Masayuki and {Taranu}, Dan S. and {Birrer}, Simon and {Huang}, Song and {Li}, Junyao and {Onodera}, Masato and {Shibuya}, Takatoshi and {Yasuda}, Naoki},
         title = "{Hyper Suprime-Cam Subaru Strategic Program: A Mass-dependent Slope of the Galaxy Size-Mass Relation at z < 1}",
-      journal = {\apj},
+      journal = {Astrophysical Journal},
      keywords = {Galaxy evolution, 594, Astrophysics - Astrophysics of Galaxies},
          year = 2021,
         month = nov,
[…]
