### 2023-09-15 #5703 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
