### 2023-09-15 #5703 @dfm
The publication page is giving a 404 which happens when the DOI link gets clicked on too quickly because of PDF caching. But, the PDF was correctly deposited so I'm going to close the issue. If the PDF doesn't display properly by tomorrow, let's check back in.

In the meantime:
